build-sys: make ./configure to seach vi path
Using "/usr/ucb/vi" as defaut fallback EDITOR is not going to work on
most of distributions where this package is installed.  That said it
might work somewhere, so searching the vi editor at configure time is the
sensible thing to do, if user does not want to define the default.

Signed-off-by: Sami Kerola <kerolasa@iki.fi>