crond: report missing newline before EOF
If the last line of a system crontab is missing a newline, it is silently
ignored.  This is because the load_env() function returns ERR when such a case
is encountered, but handling of ERR within the switch statement was erroneously
removed in 04b8ac7.

This is not a problem for user crontabs; ERR is handled there.

Steps to reproduce:

    # echo    "* * * * * root touch /tmp/foo" >  /etc/cron.d/missingeof
    # echo -n "* * * * * root touch /tmp/bar" >> /etc/cron.d/missingeof

The second job "touch /tmp/bar" is silently ignored.

To avoid this, re-introduce handling of ERR, and log an error instead so that a
sysadmin can be made aware of this problem.

Bug-Debian: https://bugs.debian.org/79037