entries: Explicitly validate upper ranges and steps
Internally, the upper end of a range and the step size are merely used as loop
conditions and increment, hence it's possible to define nonsensical entries
such as

    1-234/5678 * * * * touch /dev/null

which translates to

    for (i = 1; i <= 234; i += 5678)

Validate these numbers before using them, assuming that invalid numbers are at
least unintended and thus something we'd like to alert about. For the step
size, we further assume that the largest valid step should be no larger than
the largest possible value for the upper range (60 for minutes, etc.).

Bug-Debian: https://bugs.debian.org/533726