Fix crontab.5 man page, wrt verbiage regarding named lists/ranges for crontab entries.
The crontab.5 manpage claims "Names can also be used for the 'month' and 'day of week' fields. ... Ranges or lists of names are not allowed." Based on my reading of the source code (not entirely to be trusted) and an experiment I ran (more trustworthy), this is not actually true.

See https://github.com/cronie-crond/cronie/issues/62 for more info.